puts into one file per requirements
