+++ step1/src/example.csv
@@ -1,0 +1,4 @@
+id,type,name
+0,bird,blue
+1,cat,felix
+2,dog,toto
