+++ step1/src/transposed_example.csv
@@ -1,0 +1,3 @@
+id,0,1,2
+type,bird,cat,dog
+name,blue,felix,toto
